Rename set-up-eslint.ts
